// src/types/proyectos.ts
export type Proyecto = {
    idProyectoInmobiliario: number;
    idEmpresa: number;
    nombre: string;
    ubicacion: string;
    descripcion: string;
    fechaInicio: string;
    fechaFin: string;
    idEstadoPropiedad: number;
    telefonoContacto: string;
    emailContacto: string;
    idDistrito: number;
}